+++ code/game/AI_Mark2.cpp
@@ -61,7 +61,7 @@ void NPC_Mark2_Precache( void )
 	G_EffectIndex( "blaster/smoke_bolton" );
 	G_EffectIndex( "bryar/muzzle_flash" );
 
-	RegisterItem( FindItemForWeapon( WP_BRYAR_PISTOL ));
+	RegisterItem( FindItemForWeapon((int) WP_BRYAR_PISTOL ));
 	RegisterItem( FindItemForAmmo( 	AMMO_METAL_BOLTS));
 	RegisterItem( FindItemForAmmo( AMMO_POWERCELL ));
 	RegisterItem( FindItemForAmmo( AMMO_BLASTER ));
